Change net-snmp-devel to Suggest
net-snmp-devel require libopenssl-devel, but default on opensuse was
installed libressl. We can't install this package without resolve
this conflict. Decided to change from Requires to Suggests. If someone
want this package, must manually install and resolve this conflict.

Also change libopenssl-devel to libressl-devel. Reason is  the same as
above. libressl-devel is conflicted with libopenssl-devel.